Create a KDED module that will act as a ICC color server
The color server complies with the X Color Management specification.
http://oyranos.org/scm?p=xcolor.git;a=blob;f=docs/X_Color_Management.txt

However, it is not yet feature complete:
- Monitoring and handling of X11 events must be implemented.
- Window regions handling needs to be implemented.

Additionally, there are lots of TODO's and FIXME's still around the
code.

Heavily inspired from CompICC.
http://sourceforge.net/apps/mediawiki/compicc/index.php?title=Main_Page

Currently it is almost completely independent of the rest of
kolor-manager, but this should change in the future.