KS: fix removal of device atom
Bug:
KS omitted to set sRGB.icc profile into the _ICC_PROFILE(_xxx).
KS sets _ICC_PROFILE(_xxx) only if there is no XCM_DEVICE_PROFILE(_xxx) in
the way.

Fix:
updateProfileForAtom() now correctly looks for the base atom to
obtain the screen number and then removes the *device* atom in
XCM_DEVICE_PROFILE(_xxx). That fits the CompICC code and KS is able to set
correctly the sRGB.icc profile into the _ICC_PROFILE(_xxx).

Discussion:
Terminology was not very clear in the CompICC code. So that very likely caused
an irritation.

Fix:
The fuzzy "serverProfile" naming is corrected to "baseProfile" to remove the
reason for the assumed irritation.