Implement the transfer of the color lookup tables via shared memory
A shared memory pool is created to hold all the known color lookup
tables, in sequence. It is like a vector of the tables, in shared
memory. The color lookup tables are registered in the pool and they
become available for the clients.

The clients just need the key for the shared memory segment, and the
indices for the required color lookup tables.

This method of communication should provide a significant performance
boost, without using resources.